+++ src/main/java/me/marin/lockout/lockout/goals/kill/KillSlimeGoal.java
@@ -1,0 +1,37 @@
+package me.marin.lockout.lockout.goals.kill;
+
+import me.marin.lockout.Constants;
+import me.marin.lockout.lockout.interfaces.KillMobGoal;
+import me.marin.lockout.lockout.texture.TextureProvider;
+import net.minecraft.entity.EntityType;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.Identifier;
+
+public class KillSlimeGoal extends KillMobGoal implements TextureProvider {
+
+    public KillSlimeGoal(String id, String data) {
+        super(id, data);
+    }
+
+    @Override
+    public String getGoalName() {
+        return "Kill Slime";
+    }
+
+    @Override
+    public ItemStack getTextureItemStack() {
+        return null;
+    }
+
+    private static final Identifier TEXTURE = Identifier.of(Constants.NAMESPACE, "textures/custom/kill/kill_slime.png");
+    @Override
+    public Identifier getTextureIdentifier() {
+        return TEXTURE;
+    }
+
+    @Override
+    public EntityType<?> getEntity() {
+        return EntityType.SLIME;
+    }
+}
+