+++ src/main/java/me/marin/lockout/lockout/goals/obtain/ObtainAllDyesGoal.java
@@ -1,0 +1,45 @@
+package me.marin.lockout.lockout.goals.obtain;
+
+import me.marin.lockout.lockout.interfaces.ObtainAllItemsGoal;
+import net.minecraft.item.Item;
+import net.minecraft.item.Items;
+
+import java.util.List;
+
+public class ObtainAllDyesGoal extends ObtainAllItemsGoal {
+
+    private static final List<Item> ITEMS = List.of(
+            Items.WHITE_DYE,
+            Items.ORANGE_DYE,
+            Items.MAGENTA_DYE,
+            Items.LIGHT_BLUE_DYE,
+            Items.YELLOW_DYE,
+            Items.LIME_DYE,
+            Items.PINK_DYE,
+            Items.GRAY_DYE,
+            Items.LIGHT_GRAY_DYE,
+            Items.CYAN_DYE,
+            Items.PURPLE_DYE,
+            Items.BLUE_DYE,
+            Items.BROWN_DYE,
+            Items.GREEN_DYE,
+            Items.RED_DYE,
+            Items.BLACK_DYE
+    );
+
+    public ObtainAllDyesGoal(String id, String data) {
+        super(id, data);
+    }
+
+    @Override
+    public List<Item> getItems() {
+        return ITEMS;
+    }
+
+    @Override
+    public String getGoalName() {
+        return "Obtain all Dyes";
+    }
+
+}
+